Time: 13 ms (60.80%), Space: 17.8 MB (8.27%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -11,8 +11,8 @@
  */
 class Solution {
 public:
-    vector<long long> width;
-    long long mx;
+    vector<int> width;
+    int mx;
     void h(TreeNode *root, int depth) {
         if(!root) {
             for(int i=depth ; i<width.size() ; ++i) {
